Refactor so that each boto3 client type is in its own class
