+++ cloudwatch_monitoring/tests/test_state_machine.py
@@ -6,7 +6,7 @@ from unittest import TestCase, mock
 
 from ..positioning import Positioning
 from ..state_machine import create_state_machine_widgets
-from ..api_calls import APICalls
+from ..step_function_api_calls import StepFunctionAPICalls
 
 
 class TestCreateStateMachineWidgets(TestCase):
@@ -90,11 +90,11 @@ class TestCreateStateMachineWidgets(TestCase):
             ]
         }
 
-    @mock.patch('cloudwatch_monitoring.api_calls.boto3.client', autospec=True)
+    @mock.patch('cloudwatch_monitoring.step_function_api_calls.boto3.client', autospec=True)
     def test_get_all_state_machines(self, m_client):
         mock_stepfunctions_client = mock.Mock()
         m_client.return_value = mock_stepfunctions_client
-        api_calls = APICalls(self.region, self.client_type, self.deploy_stage)
+        api_calls = StepFunctionAPICalls(self.region, self.deploy_stage)
 
         # only one state machine returned from list_state_machines
         mock_stepfunctions_client.list_state_machines.return_value = self.state_machine_list_no_next_token
@@ -111,11 +111,11 @@ class TestCreateStateMachineWidgets(TestCase):
         # assert the stepfunctions client called list_state_machines with expected arguments
         mock_stepfunctions_client.list_state_machines.assert_called_with(maxResults=self.max_results)
 
-    @mock.patch('cloudwatch_monitoring.api_calls.boto3.client', autospec=True)
+    @mock.patch('cloudwatch_monitoring.step_function_api_calls.boto3.client', autospec=True)
     def test_get_all_state_machines_next_token_pagination(self, m_client):
         mock_stepfunctions_client = mock.Mock()
         m_client.return_value = mock_stepfunctions_client
-        api_calls = APICalls(self.region, self.client_type, self.deploy_stage)
+        api_calls = StepFunctionAPICalls(self.region, self.deploy_stage)
 
         # 2 state machines returned, the first state machine has the key nextToken that causes us to begin iterating 
         # through pages of list_state_machines responses
@@ -141,12 +141,12 @@ class TestCreateStateMachineWidgets(TestCase):
         # assert the list_state_machines calls were called, and in the expected order
         mock_stepfunctions_client.list_state_machines.assert_has_calls(expected, any_order=False)
 
-    @mock.patch('cloudwatch_monitoring.api_calls.boto3.client', autospec=True)
+    @mock.patch('cloudwatch_monitoring.step_function_api_calls.boto3.client', autospec=True)
     def test_is_iow_state_machine_filter_happy_path(self, m_client):
         mock_stepfunctions_client = mock.Mock()
         m_client.return_value = mock_stepfunctions_client
         mock_stepfunctions_client.list_tags_for_resource.return_value = self.tags_list_for_valid_state_machine_arn_1
-        api_calls = APICalls(self.region, self.client_type, self.deploy_stage)
+        api_calls = StepFunctionAPICalls(self.region, self.deploy_stage)
 
         # assert the return value is true, since list_tags_for_resource returned a valid response
         # noinspection PyPackageRequirements
@@ -160,12 +160,12 @@ class TestCreateStateMachineWidgets(TestCase):
         # assert the stepfunctions client called list_tags_for_resource with expected arguments
         mock_stepfunctions_client.list_tags_for_resource.assert_called_with(resourceArn=self.valid_state_machine_arn_1)
 
-    @mock.patch('cloudwatch_monitoring.api_calls.boto3.client', autospec=True)
+    @mock.patch('cloudwatch_monitoring.step_function_api_calls.boto3.client', autospec=True)
     def test_is_iow_state_machine_filter_empty_tags_response(self, m_client):
         mock_stepfunctions_client = mock.Mock()
         m_client.return_value = mock_stepfunctions_client
         mock_stepfunctions_client.list_tags_for_resource.return_value = self.tags_list_empty_tags
-        api_calls = APICalls(self.region, self.client_type, self.deploy_stage)
+        api_calls = StepFunctionAPICalls(self.region, self.deploy_stage)
 
         # assert the return value is False, since list_tags_for_resource returned a response with no 'Tags' key
         # noinspection PyPackageRequirements
@@ -173,12 +173,12 @@ class TestCreateStateMachineWidgets(TestCase):
             api_calls.is_iow_state_machine_filter(self.valid_state_machine_arn_1)
         )
 
-    @mock.patch('cloudwatch_monitoring.api_calls.boto3.client', autospec=True)
+    @mock.patch('cloudwatch_monitoring.step_function_api_calls.boto3.client', autospec=True)
     def test_is_iow_state_machine_filter_no_tags_response(self, m_client):
         mock_stepfunctions_client = mock.Mock()
         m_client.return_value = mock_stepfunctions_client
         mock_stepfunctions_client.list_tags_for_resource.return_value = self.tags_list_no_tags
-        api_calls = APICalls(self.region, self.client_type, self.deploy_stage)
+        api_calls = StepFunctionAPICalls(self.region, self.deploy_stage)
 
         # assert the return value is False, since list_tags_for_resource returned a response with no
         # tags in the Tags dict
@@ -187,12 +187,12 @@ class TestCreateStateMachineWidgets(TestCase):
             api_calls.is_iow_state_machine_filter(self.valid_state_machine_arn_1)
         )
 
-    @mock.patch('cloudwatch_monitoring.api_calls.boto3.client', autospec=True)
+    @mock.patch('cloudwatch_monitoring.step_function_api_calls.boto3.client', autospec=True)
     def test_is_iow_state_machine_filter_no_wma_org_key_response(self, m_client):
         mock_stepfunctions_client = mock.Mock()
         m_client.return_value = mock_stepfunctions_client
         mock_stepfunctions_client.list_tags_for_resource.return_value = self.tags_list_no_wma_org_key
-        api_calls = APICalls(self.region, self.client_type, self.deploy_stage)
+        api_calls = StepFunctionAPICalls(self.region, self.deploy_stage)
 
         # assert the return value is False, since list_tags_for_resource returned a response with no
         # wma:organization key in the Tags dict
@@ -202,12 +202,12 @@ class TestCreateStateMachineWidgets(TestCase):
             api_calls.is_iow_state_machine_filter(self.valid_state_machine_arn_1)
         )
 
-    @mock.patch('cloudwatch_monitoring.api_calls.boto3.client', autospec=True)
+    @mock.patch('cloudwatch_monitoring.step_function_api_calls.boto3.client', autospec=True)
     def test_is_iow_state_machine_filter_no_iow_value_response(self, m_client):
         mock_stepfunctions_client = mock.Mock()
         m_client.return_value = mock_stepfunctions_client
         mock_stepfunctions_client.list_tags_for_resource.return_value = self.tags_list_no_iow_value
-        api_calls = APICalls(self.region, self.client_type, self.deploy_stage)
+        api_calls = StepFunctionAPICalls(self.region, self.deploy_stage)
 
         # assert the return value is False, since list_tags_for_resource returned a response with a valid
         # wma:organization key, but the value is not "IOW"
@@ -216,7 +216,7 @@ class TestCreateStateMachineWidgets(TestCase):
             api_calls.is_iow_state_machine_filter(self.valid_state_machine_arn_1)
         )
 
-    @mock.patch('cloudwatch_monitoring.state_machine.APICalls', autospec=True)
+    @mock.patch('cloudwatch_monitoring.state_machine.StepFunctionAPICalls', autospec=True)
     def test_create_state_machine_widgets(self, m_api_calls):
         # return values
         m_api_calls.return_value.get_all_state_machines.return_value = self.state_machine_list_after_successful_pagination
